COMP: Address std::cout.precision conversion warning.
Address conversion warning on Linux-x86_64-gcc4.1 (lhcp-vm-rh502.nlm) when
setting the output stream precision.

+++ Modules/Filtering/ImageIntensity/test/itkSquareImageFilterTest.cxx
@@ -122,7 +122,7 @@ int itkSquareImageFilterTest( int, char* [] )
             static_cast<OutputImageType::PixelType>( x2 );
     if( !itk::Math::FloatAlmostEqual( square, output, 10, epsilon ) )
       {
-      std::cerr.precision( itk::Math::abs( std::log10( epsilon ) ) );
+      std::cerr.precision( unsigned( itk::Math::abs( std::log10( epsilon ) ) ) );
       std::cerr << "Error in itkSquareImageFilterTest " << std::endl;
       std::cerr << " square( " << input << ") = " << square << std::endl;
       std::cerr << " differs from " << output;
